pyproject.toml
[project]
name = "mas-v0.2"
version = "0.1.0"
description = "Multi-Agent System for Digital Marketing"
authors = [
    {name = "HuanMu Team", email = "consumerclone@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "anthropic>=0.60.0",
    "elasticsearch>=9.1.0",
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.0",
    "langchain>=0.3.27",
    "langgraph>=0.6.2",
    "openai>=1.98.0",
    "pillow>=10.0.0",
    "pydantic-settings>=2.10.1",
    "pymilvus>=2.5.14",
    "python-multipart>=0.0.6",
    "redis>=6.2.0",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "black>=25.1.0",
    "isort>=5.14.0",
]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
